refactor: :card_file_box: Refactorizacion para arquitectura CLEAN
Se refactorizó el código y la estructura de carpetas/paquetes, para representar la arquitectura CLEAN para Flutter, la cual se asemeja a la arquitectura hezagonal utilizada en APIs de backend.
