+++ lib/features/product/data/product_fetch.dart
@@ -1,0 +1,59 @@
+import 'dart:async';
+import 'dart:convert';
+import 'package:flutter/material.dart';
+import 'package:godeliveryapp_naranja/features/product/domain/product.dart';
+import 'package:godeliveryapp_naranja/features/product/presentation/widgets/product_card.dart';
+import 'package:http/http.dart' as http;
+
+Future <List<Product>> fetchProducts() async {
+  final response = await http.get(Uri.parse('https://orangeteam-deliverybackend-production.up.railway.app/product'));
+
+  print(response.body);
+  print(response.statusCode);
+  if (response.statusCode == 200) {
+    // If the server did return a 200 OK response,
+    // then parse the JSON.
+    final Map<String,dynamic> jsonData = jsonDecode(response.body);
+    final List<dynamic> productsData = jsonData['products'];
+    return productsData.map((productJson) => Product.fromJson(productJson)).toList();
+  } else {
+    // If the server did not return a 200 OK response,
+    // then throw an exception.
+    throw Exception('Failed to fetch products');
+  }
+}
+
+class ProductListScreen extends StatefulWidget {
+  const ProductListScreen({super.key});
+
+  @override
+  State<ProductListScreen> createState() => _ProductListScreenState();
+}
+
+class _ProductListScreenState extends State<ProductListScreen> {
+  late Future<List<Product>> futureProducts;
+
+  @override
+  void initState() {
+    super.initState();
+    futureProducts = fetchProducts();
+    print(futureProducts);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return FutureBuilder<List<Product>>(
+      future: futureProducts,
+      builder: (context, snapshot) {
+        if (snapshot.hasData){
+          return Wrap(
+        children: snapshot.data!.map((product) => ProductItem(product: product)).toList(),
+      );
+        } else if (snapshot.hasError){
+          return Text('${snapshot.error}');
+        }
+        return const Center(child: CircularProgressIndicator(color: Colors.orange,));
+      }
+    );
+  }
+}
