+++ lib/features/product/domain/product.dart
