--- lib/orderhistory/order_history_screen.dart
@@ -1,50 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:godeliveryapp_naranja/orderhistory/navbar.dart';
-
-import 'custom_tab_bar.dart';
-
-class OrderHistoryScreen extends StatefulWidget {
-  const OrderHistoryScreen({super.key});
-
-  @override
-  _OrderHistoryScreenState createState() => _OrderHistoryScreenState();
-}
-
-class _OrderHistoryScreenState extends State<OrderHistoryScreen>
-    with SingleTickerProviderStateMixin {
-  late TabController _tabController;
-  int _selectedIndex = 0;
-
-  @override
-  void initState() {
-    super.initState();
-    _tabController = TabController(length: 2, vsync: this);
-  }
-
-  @override
-  void dispose() {
-    _tabController.dispose();
-    super.dispose();
-  }
-
-  void _onNavBarTapped(int index) {
-    setState(() {
-      _selectedIndex = index;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Historial de ordenes'),
-        centerTitle: true,
-      ),
-      body: CustomTabBar(tabController: _tabController),
-      bottomNavigationBar: CustomNavBar(
-        currentIndex: _selectedIndex,
-        onTap: _onNavBarTapped,
-      ),
-    );
-  }
-}
