--- lib/product_detail.dart
@@ -1,221 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:godeliveryapp_naranja/orderhistory/navbar.dart';
-import 'package:godeliveryapp_naranja/carrito/cart_screen.dart';
-import 'package:godeliveryapp_naranja/presentation/interfaces/loading_screen.dart';
-
-
-class ProductDetailScreen extends StatefulWidget {
-  const ProductDetailScreen({super.key});
-
-  @override
-  _ProductDetailScreenState createState() => _ProductDetailScreenState();
-}
-
-class _ProductDetailScreenState extends State<ProductDetailScreen> {
-  int quantity = 1;
-  double price = 25.0;
-  final String productName = "Nestle Koko Krunch Duo (Kids Pack)";
-  final String descripcion =
-      "Delicioso cereal de chocolate para niños, perfecto para el desayuno o la merienda.";
-  final String divisa = "USD"; // Divisa del precio
-
-  int _currentIndex = 0; // Variable para el índice de la barra de navegación
-
-  // Función para manejar el cambio de índice en el navbar
-  void _onTap(int index) {
-    setState(() {
-      _currentIndex = index;
-    });
-  }
-
-  void incrementQuantity() {
-    setState(() {
-      quantity++;
-      price = 25.0 * quantity;
-    });
-  }
-
-  void decrementQuantity() {
-    setState(() {
-      if (quantity > 1) {
-        quantity--;
-        price = 25.0 * quantity;
-      }
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        centerTitle: true,
-        title: const Text(
-          'Detalle Producto',
-          style: TextStyle(
-            color: Color.fromARGB(255, 175, 91, 7),
-            fontSize: 24,
-            fontWeight: FontWeight.bold,
-          ),
-        ),
-        leading: IconButton(
-          icon: const Icon(Icons.arrow_back,
-              color: Color.fromARGB(255, 175, 91, 7)),
-          onPressed: () {
-            Navigator.pop(context);
-          },
-        ),
-        actions: [
-          IconButton(
-            icon: const Icon(Icons.shopping_cart,
-                color: Color.fromARGB(255, 175, 91, 7)),
-            onPressed: () {
-              showLoadingScreen(context, destination: const CartScreen());
-            },
-          ),
-        ],
-      ),
-      body: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: [
-            Container(
-              padding: const EdgeInsets.all(16),
-              decoration: BoxDecoration(
-                color: Colors.white,
-                borderRadius: BorderRadius.circular(12),
-                boxShadow: const [
-                  BoxShadow(
-                    color: Colors.black12,
-                    blurRadius: 8,
-                    offset: Offset(0, 4),
-                  ),
-                ],
-              ),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Center(
-                    child: Image.network(
-                      'https://media.nedigital.sg/fairprice/fpol/media/images/product/XL/13206550_XL1_20240903.jpg?w=400&q=70',
-                      height: 150,
-                    ),
-                  ),
-                  const SizedBox(height: 18),
-                  Center(
-                    child: Text(
-                      productName,
-                      style: const TextStyle(
-                          fontSize: 20, fontWeight: FontWeight.bold),
-                    ),
-                  ),
-                  const SizedBox(height: 8),
-                  Center(
-                    child: Text(
-                      descripcion,
-                      style:
-                          const TextStyle(fontSize: 16, color: Colors.black54),
-                      textAlign: TextAlign.center,
-                    ),
-                  ),
-                  const SizedBox(height: 8),
-                  Divider(color: Colors.grey[300]),
-                  const SizedBox(height: 8),
-                  const Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      Text(
-                        'Presentación',
-                        style: TextStyle(fontSize: 16, color: Colors.black54),
-                      ),
-                      Text(
-                        '550 gm',
-                        style: TextStyle(fontSize: 16, color: Colors.black87),
-                      ),
-                    ],
-                  ),
-                  const SizedBox(height: 8),
-                  Divider(color: Colors.grey[300]),
-                  const SizedBox(height: 8),
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      const Text(
-                        'Precio',
-                        style: TextStyle(fontSize: 16, color: Colors.black54),
-                      ),
-                      Text(
-                        '$divisa $price', // Añadimos la divisa junto con el precio
-                        style: const TextStyle(
-                          fontSize: 22,
-                          color: Color(0xFFFF9027),
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                    ],
-                  ),
-                ],
-              ),
-            ),
-            const Spacer(),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                ElevatedButton(
-                  onPressed: decrementQuantity,
-                  style: ElevatedButton.styleFrom(
-                    shape: const CircleBorder(),
-                    backgroundColor: const Color.fromARGB(255, 175, 91, 7),
-                    padding: const EdgeInsets.all(12),
-                  ),
-                  child: const Icon(Icons.remove, color: Colors.white),
-                ),
-                Padding(
-                  padding: const EdgeInsets.symmetric(horizontal: 20),
-                  child: Text(
-                    '$quantity',
-                    style: const TextStyle(fontSize: 20),
-                  ),
-                ),
-                ElevatedButton(
-                  onPressed: incrementQuantity,
-                  style: ElevatedButton.styleFrom(
-                    shape: const CircleBorder(),
-                    backgroundColor: const Color.fromARGB(255, 175, 91, 7),
-                    padding: const EdgeInsets.all(12),
-                  ),
-                  child: const Icon(Icons.add, color: Colors.white),
-                ),
-                const SizedBox(width: 16),
-                Expanded(
-                  child: ElevatedButton.icon(
-                    onPressed: () {},
-                    icon: const Icon(Icons.shopping_cart, color: Colors.white),
-                    label: const Text(
-                      'Añadir al carrito',
-                      style: TextStyle(color: Colors.white, fontSize: 14),
-                    ),
-                    style: ElevatedButton.styleFrom(
-                      backgroundColor: const Color(0xFFFF7000),
-                      padding: const EdgeInsets.symmetric(vertical: 14),
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(8),
-                      ),
-                      textStyle: const TextStyle(fontSize: 18),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(height: 16),
-          ],
-        ),
-      ),
-      // Agregar CustomNavBar en el bottomNavigationBar
-      bottomNavigationBar: CustomNavBar(
-        currentIndex: _currentIndex,
-        onTap: _onTap,
-      ),
-    );
-  }
-}
